requirements.txt
certifi==2017.4.17
chardet==3.0.4
future==0.16.0
idna==2.5
lxml==3.8.0
pkg-resources==0.0.0
py==1.4.34
pytest==3.1.3
pytest-benchmark==3.0.0
python-telegram-bot==6.1.0
requests==2.18.1
urllib3==1.21.1
